Avoid concurrent reading of the notify FIFO file
When reloading a configuration with the [vrrp_]notify_fifo_script
option, the current FIFO notify script is killed with TERM signal and
then a new instance of script is started. If the [vrrp_]notify_fifo file
is unchanged, both instances of the script has read access to the same
FIFO during 500ms due to the following in the
doc/samples/sample_notify_fifo.sh

> ( sleep 0.5
>  kill -ALRM $$
> ) &

As a consequence, the FIFO is sometimes read at the same time by both
scripts that get corrupted input.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>