fifo: In sample_notify_fifo.sh redirect stderr for kill command
If the shell that is running the main fifo script terminates while
the subshell is executing sleep 0.5, then $$ (the main script) may have
terminated before the kill command is executed in the subshell. Since
there is no problem if that happens, and it is quite legitimate, redirect
the error output of kill.

Many thanks to Louis Scalbert <louis.scalbert@6wind.com> for noticing this.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>