fifo: avoid concurrent instances of sample_notify_fifo.sh
When keepalived reloads its configuration, the notify FIFO script is
restarted. The script that was running before the reload does not end
immediately. Then the new one can run concurrently.

Note that although the ALRM trap is scheduled 500ms after the shutdown,
if a command is running when the 500ms period is over, the script waits
for the command end before calling stopping().

Use a PID file to avoid running concurrent instances of
sample_notify_fifo.sh. If the script exists without removing the PID
file, the new instance checks that no file descriptor is open for the
PID file with the PID that it contains. Checking that the PID does not
exist in the process list is not sufficient as a PID may be reused by
another process.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>