fifo: avoid two instances of sample_notify_fifo.sh reading FIFO
At reload, the notify FIFO script is restarted.

There is a theoretical risk of corruption of the FIFO input. Since the
notify FIFO script running before the reload and the one running
afterwards both read to the same FIFO file, and if there is a delay in
scheduling the originally executing FIFO script, it is possible that
both instances of the script read the FIFO file at the same time.

This patch use a PID file to make sure that there is no concurrent
reading of the FIFO file.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>