vrrp: Don't assume scripts added at reload are in failed state
If a vrrp_script was added during a reload, keepalived put it
initially into failed state, since that is what happens when
keepalvied states. The reason for putting vrrp_scripts initially
into failed states is to stop vrrp instances becoming master and
then almost immediately going to failed state when a vrrp_script
completes for the first time.

The essential idea is to stop unnecessary state transitions of
vrrp instances due to the initial status of a script not being
known.

In order to follow the idea for adding scripts during a reload, we
cannot default to a script being either succesful or failed, since
either may cause the priority of an instance to change, or cause a
vrrp instance in master state to transition to fault state until
the script completes.

This commit makes keepalived not apply any result for a script added
at a reload until the script actually completes. This stops any
short term effect based on an assumption of the state of the script
before it actually returns a result. This means a script can be
added at a reload, and a vrrp instance in master state will not
transition to fault state and then to backup followed by master
state if the script returns success, but rather it will remain in
master state. If the script returns failure, then the vrrp instance
will transition to fault state once the script returns failure.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>