all: Ensure PDEATHSIG is not disabled
When keepalived checks whether scripts can be accessed, it executes
seteuid/setegid if the script has a user/group configured. Calling
these functions clears PDEATHSIG, so it needs to be restore after
resetting the euid/egid.

There are also windows when PDEATHSIG is not set, i.e. as above and
between fork() and setting PDEATHSIG, and during the window the parent
process may terminate and so the child processes' parent process
changes. We now record the parent PID before forking, and after
PDEATHSIG is set, check that we still have the same parent PID.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>