vrrp: Add add/prepend/append options to static and virtual routes
The kernel by default prepends routes, whereas the ip (iproute2)
utility be default adds routes (adding a route does not allow duplicates
whereas appending or prepending does). keepalived previously has not
set the flags relating to this, and so has always prepended routes. This
means that duplicate routes could be created.

This commit adds the 'add', 'prepend' (default) and 'append' options for
routes.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>