doc: update pre-requisites installation
in rhel based OS prerequisites section, it says to install `automake`
but I was unable to build keepalived from source without installing `make` instead.

Also, `ipset-devel`, `libnfnetlink-devel` and `file-devel` packages require the repo
`PowerTools` to be enabled so I thought it was worth it to mention how to install them.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>