vrrp: Add support for Infiniband over IPv6
Github issue #2100 reported that attempting to use IPv6	over Infinband
was causing keepalived to segfault. It turned out that vrrp_ndisc.c
had a comment that it still needed to be implemented, which we have now
been able to do with someone in a position to test it.

With many thanks for Itel Levy of NVIDIA, Israel for reporting the
issue and and testing the patch to confirm that it works.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>