config parser: Fix segfault caused by extra '}' and other parser fixes
If there was a configuration error in a block, e.g. a vrrp_instance,
keepalived would apply the configuration in the rest of the block to
the previous object of that type, e.g. the previous vrrp instance. If
there had been no previous instance, keepalived would probably segfault.

This commit changes the way the parser works. A new instance of an
object, e.g. a VRRP instance or a virtual server, is only added to the
list of those objects once the configuration of that object is complete.
In particular it no longer applies the configuration to the last entry
on the list of the relevant object type, but keeps a point to the
object currently being configured.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>