check: use last entry if duplicate definition
Commits 8a3f145 - "fix mem leaks when virtualhost and snmp_name
are duplicate" and 86bbb2e - "fix mem leaks when virtualhost is
duplicate" changed the behaviour of virtualhost and snmp_name
configuration if there were duplicate definitions to use the
first defined entry, whereas previously the last defined entry
was used, albeit with a memory leak.

This commit retains the memory leak fixes, but reverts the
behaviour to use the last definition rather than the first.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>