config_check: Don't kill process when checking validity of scripts
When a configuration is checked keepalived also checks script security
and calls check_script_secure().

The same function is also called during normal keepalived lifecycle at
startup. The function has a check that the main pid didn't die while
we were checking other things (see #5927172 for details).

main_pid is set in start_keepalived() function called from
keepalived_main(). In keepalived_main() we parse args, decide if we
have to test configuration and do the check before calling
start_keepalived(). So main_pid is not set to a proper value.

The best check is !test_bit(CONFIG_TEST_BIT, &debug) since that
identifies the reason for not checking the pid.

Many thanks to github user oxpa for identifying the issue and
providing the original patch.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>