vrrp: Work around a GCC LTO bug in checksum calculation
There appears to be a GCC LTO bug, which caused in_csum function, when
built with LTO, to return incorrect results.

The problem has been observed with GCC versions 11.2, 11.3.1 and 12.1.1,
on Ubuntu 22.04, Fedora 34, Fedora 36 and Fedora 37. No working versions
of GCC have been identified.

The problem does not occur when not using LTO, nor when using clang, even
with LTO.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>