malloc: add option to check for malloc's etc returning NULL
Configure option --enable-malloc-check will cause the returned
value of malloc/realloc/strdup/strndup to be checked to ensure
that they do not return NULL. If any such call does return NULL
a message will be logged and the process will terminate.

Unless sysctl vm.overcommit_memory == 2 (default is usually 0),
or the malloc would cause the process virtual address space to
exceed the limit, malloc etc will not return NULL. It is only
once there is a write into the memory block that the memory is
actually allocated, and if there is insufficient memory (including
swap space), then the OOM killer will step in to either kill
keepalived, or kill another process. Consequently checking for
NULL being returned is generally a waste of time and program size.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>