netlink: do not ignore vip as system owner
When a VRRP instance is waiting for an IP address on the interface to
start, if the vip is added on the interface, keepalived ignores it and
remains in FAULT state, even if the VRRP instance is the system owner
(priority 255 and vip = system IP). However, when adding an IP address
different than the vip, the instance can change the state.

This makes nonsense for an instance is configured as the system owner
transit to FAULT state since the system has not yet the vip.
When the vip is added to the system, keepalived ignores it
and this instance remains in FAULT state, whereas an IP address
different than the vip can change the state of this instance.

Thus, do not ignore if a vip is added to the system when a VRRP
instance is configured with 255 priority.

Signed-off-by: Loïc Sang <loic.sang@6wind.com>