VRRP: add v3_checksum_as_v2 configuration option
RFC 5798 (the VRRPv3 RFC) states regarging the checksum:

5.2.8.  Checksum

   The checksum field is used to detect data corruption in the VRRP
   message.

   The checksum is the 16-bit one's complement of the one's complement
   sum of the entire VRRP message starting with the version field and a
   "pseudo-header" as defined in Section 8.1 of [RFC2460].  The next
   header field in the "pseudo-header" should be set to 112 (decimal)
   for VRRP.  For computing the checksum, the checksum field is set to
   zero.  See RFC1071 for more detail [RFC1071].

Some manufacturers (e.g. Cisco) interpret this to mean that the pseudo-
header is not included in the checksum calculation, since RFC2460 only
defines a pseudo-header for IPv6. RFC3768 (the last VRRPv2 RFC) did not
include a pseudo-header in the checksum. However, keepalived has always
included a pseudo-header in the VRRPv3 IPv4 checksum, which is also
consistent with the default setting in Wireshark.

In order to allow interoperation with Cisco routers, and possibly other
manufacturers, the "v3_checksum_as_v2" keyword, when configured in
global_defs to set the default for all vrrp_instances, or in individual
vrrp_instances, causes those vrrp_instances to exclude the pseudo-
header from the checksum. The default action of including the pseudo-
header in the checksum remains unchanged.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>