vrrp: prepare options for vrrp system owner instance
Add a global option "vrrp_system_owner" to define the
behavior of a vrrp configured with 255 priority (system owner):

- The VRRP_FLAG_SYSTEM_OWNER_DFT is the current implementation.
A system owner instance can transit with any IP address on the physical
interafce. This is the default option when vrrp_system_owner is not defined.

- With VRRP_FLAG_SYSTEM_OWNER_ANY, a system owner instance can transit
with any VIP configured on the physical interface. To use, that define
"vrrp_system_owner any" in keepalived.conf.

- With VRRP_FLAG_SYSTEM_OWNER_STRICT, a system owner instance can
transit when all VIPs are configured on the physical interface.
To use that, define "vrrp_system_owner strict" in keepalived.conf.

Signed-off-by: Loïc Sang <loic.sang@6wind.com>