vrrp: enhance vrrp instance with priority 255
Add two more options for a vrrp instance with 255 priority (system owner)
i) Remain in fault state until all the VIPs have been added.
For that, set "vrrp_system_owner strict" in keepalived.conf.

ii) If any of the VIPs exists, add all the remaining VIPs.
For that, set "vrrp_system_owner any" in keepalived.conf.

iii) Go to MASTER state if there is an IP address on the VRRP instance's
interface (this it current implemenation).
This is the default behaviour, if vrrp_system_owner is not configured.

Signed-off-by: Loïc Sang <loic.sang@6wind.com>