vrrp: Add option to revert to backup if thread timer expires
If the VRRP process is not scheduled for sufficiently long, another VRRP
instance may have taken over as master. For some users, minimising the
number of master switches is desired, and so if nopreempt is configured
(if it is not configured the highest priority instance will take over as
master again), and if it is too long after a thread timer expires before
keepalived is scheduled to run so that another instance will probably
have taken over as master, we will just revert to backup state rather
than sending further adverts.

The keyword that configures this is thread_timer_expired.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>