vrrp: work around missing promiscuous netlink notifications
If the base interface does not implement IFF_UNICAST_FLT, for example
it is a bridge interface, no netlink notification is sent by the kernel
when promiscuity is set on the base interface.

The promiscuous state of the base interface is correct in kernel but it
is in incorrect in daemons that listen to the interface netlink messages
(eg. DPDK).

The issue is still there in kernel 5.4.0

Force a notification by re-setting IFLA_GROUP for the base interface.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>