vrrp: restore the vmac ipv6 link-local after flapping
The user is not supposed to shutdown a vmac interface created by
keepalived. However, it can mistakenly happen. When the link is
re-established, the link-local has disappear (the kernel removes all
IPv6 addresses on link down except if keep_addr_on_down sysctl is on)
and sending VRRP packet is no nore possible.

Restore the IPv6 Link-Local after a VMAC interface flapping. A
Link-Local is not set when the VRRP packets are sent from the base
interface (vmac-xmit-base).

Note that the IPv6 Virtual Addresses are also removed on link down which
is the desired behavior. Enabling keep_addr_on_down sysctl would keep
the link-local without this patch but would break this behavior.

Original patch signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>
Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>