vrrp: fix persistent FAULT state with use_vmac on a down base interface
If the base interface of VMAC is down at startup, the VRRP instance will
remain on FAULT state even if the base interface is set up.

keepalived uses a vrrp->num_script_if_fault fault counter that
is incremented when there is a new fault and decremented when the fault
is recovered. When the counter equals to zero, it considers there is no
fault. try_up_instance() is used to decrease the counter. Since
4c7a94a409 ("vrrp: Make VMAC IPv6 link local address mirror parent
interface"), it was not called anymore when a IPv6 link-local was back
on the VMAC interface.

Increment vrrp->num_script_if_fault when an IPv6 appears on the VMAC.

Fixes: 4c7a94a409 ("vrrp: Make VMAC IPv6 link local address mirror parent interface")
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>