vrrp: fix non-existent vmac base interface at reload
When an existing VMAC VRRP instance is updated with a base interface
that does not exists, the instance goes to FAULT state for a good
reason. However, when the base interface is appears, the VMAC interface
is never updated so the instance remains in FAULT state.

Delete the VMAC interface if the base interface is changed to a
non-existent one. The VMAC interface will be re-created when the base
interface appears.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>