vrrp: use instance fault flags instead of a counter
The num_script_if_fault counter is not very convenient. We don't easily
know what has incremented and decremented the counter. And the counter
could go back to zero for a reason in case of a bug.

Use an instance fault flags 'flags_if_fault' instead of a counter. Keep
the a counter 'num_track_fault' for tracked object.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>