ipvs: Remove checkers_queue
A configuration with 2277 virtual servers, with a total of 37205
real servers with each real server having one checker was taking
132 seconds to reload. This commit reduces the reload time to
0.24 seconds, a reduction of 99.8%!

The problem was due to every real server iterating through all
checkers, 37205 * 37205 = 1,384,212,025 iterations, not only once
but several times.

The code now maintains a list of checkers for each real server. The
disadvantage of this is that to iterate through all checkers requires
iterating through all virtual servers, and all their real servers and
then for each real server the list of checkers. If there are
relatively few checkers compared to real servers, this will take longer
than using the checkers_queue, but using a queue per real server is
still fast, and the only time the code iterates through all the checkers
is at startup/reload, other than dumping the configuration.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>