vrrp: Stop link local VMAC address responging to neighbour solicit
When an IPv6 VRRP instance using  VMAC is in backup state, the
link local address configured on the VMAC interface is the same as
the link local address on the parent interface of the VMAC. This
causes a problem with switches learning the MAC address of the
VMAC is now on the backup. This causes packets meant to be sent to
the master being sent to the backup.

This commit uses nftables/iptables to stop neighbour advertisements
for the link local address of the VMAC interface and its parent
interface being sent from the VMAC interface.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>