vrrp: work around missing promiscuous netlink notifications
If the base interface does not implement IFF_UNICAST_FLT, for example
it is a bridge interface, no netlink notification is sent by the kernel
when promiscuity is set on the base interface.

The promiscuous state of the base interface is correct in the kernel but
it is in incorrect in daemons that listen to the interface netlink
messages (eg. DPDK).

The issue is still there in kernel 6.4.6.

Force a notification by re-setting IFLA_GROUP for the base interface.

Initial patch by and signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>