build: Add configure option --enable-reproducable-build
There are users who want reproducable builds in the event that
the paths to header files change between builds.

Due to 'keepalived -v' outputting (build) configure options, if
the location of the header files changed, then the keepalived
binary would change.

If keepalived is build with the --enable-reproducable-build option,
then the configure options are written to
keepalived/keepalived.config-opts in the build tree, which is then
installed to $(DESTDIR)/etc/keepalived/keepalived.config-opts (or
whereever the default config file is installed).

keepalived reads this file for generating the 'keepalived -v'
output (and also coredump log entries), and so this file must be
installed correctly for keepalived to run.

If --enable-reproducable-build is not specified to configure, then
the behaviour of keepalived and the required installed files are
not changed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>