ipvs: Reinstate non-failed real servers if remove sorry server
When there is no sorry server, the quorum is not used, and real
servers are only removed if a checker fails. On the other hand
if there is a sorry server, if the number of alive real servers
falls below the quorum, all non-failed real servers are removed
when the sorry server is added. If the sorry server is remomed
from the configuration, non-failed real servers need to be
reinstated.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>