build: don't check or include some kernel headers if inappropriate for build
Some kernel headers are not needed if building without IPVS support, and
similarly some some kernel headers are not needed if building without VRRP
support.

In these cases, do not check for the existence of some of the headers at
configure time, and do not include the headers at compile  time.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>