all: add global keyword use_symlink_paths
By default keepalived resolves all symbolic links in path names
of scripts to the real path. This commit adds the use_symlink_paths
option to maintain the symlinks in paths, so that users can update
symlinks in order to update the scripts being called.

See github.com keepalived issue #2393 for the rational for this
requirement.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>