ipvs: Update status code of misc checker if changes while in fault state
The exit code of a misc checker can be read via SNMP. The misc check code
was not updating the last exit code if the checker was not dynamic, the
checker was already down (i.e. returned a non 0 exit code), and the exit
code changed from the previous exit code. This meant that the exit code
reported via SNMP was not the latest exit code, but the exit code that
caused the status of the checker to change.

This commit now updates the last exit code, even if the checker is already
down.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>