vrrp: remove need for route to have configured interface to track it
If a virtual route did not have an interface configured, keepalived would
log a warning saying that it could not track the route, and then would
disable tracking of that route.

It appears that it is not necessary to know the interface in order to track
the route, and in any event the netlink message received after adding the
route identifies the interface for the route if it is appropriate.

So this commit removes the requirement to specify an interface in order to
track a route.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>