vrrp: Allow for Ethernet frame padding for short packets
Some network interface cards do not strip Ethernet frame padding
before passing a packet to userspace (recvmesg()). keepalived
checks the received packet length but wasn't allowing for extra
bytes to be received that were added as frame padding.

This commit allows for frame padding to be received and not report
an incorrect packet length.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>