core: ensure only one instance of keepalived can run per config_id
There was a window when keepalived starts up when if two (or more) instances
were starting at the same time, they might not detect the other instance is
running.

This commit add advisort file locking on the PID files to ensure that only one
instance can run at a time.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>