vrrp ipvs: Stop setting SO_LINGER on TCP sockets
Setting SO_LINGER causes the close() call to block until the first
of:
1. the ACK of the FIN is received
2. the SO_LINGER timeout expires

Since the SO_LINGER timeout was set to 5 seconds, if the FIN or the
subsequent ACK were lost, then keepalived would block for 5 seconds,
which must not be allowed to happen.

The only TCP sockets that keepalived opens are for TCP_CHECK, HTTP_GET,
SSL_GET, SMTP_CHECK and sending notify emails. For all of these, for
any data that keepalived sends it receives data in response, and so
there is no purpose in using SO_LINGER.

Removing setting SO_LINGER will stop the occasional 'A thread timer
expired 5.1nnnnn seconds ago', as reported in issue #2271.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>