lib: add micro-second timers to memory allocation debugging
Previously the time was logged for memory allocation/freeing operations
in seconds. When comparing when memory was allocated/freed to debugging
logging via a log file, it was helpful, in terms of being able to
identify the sequence of events, to have the time of memory allocations
etc logged in micro-seconds.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>