lib: fix corruption of master-child_pid red black tree
Child process thread_t structures use two red-black trees, one for
the timeout, and the other for pids. It is important to ensure that
threads are removed from the child_pid RB tree at the correct time.
This was not happening when reloads were occurring and there was a
THREAD_CHILD_TIMEOUT thread on the ready list.

A few other instances of the thread not being removed from the
child_pid RB tree correctly, which are also resolved by this commit.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>