vrrp: on reload only configured track_script name was checked
On a reload, only the configured name of a track_script was being
checked to see if the new config track_script matched the old config
track_script. If the script to be executed were changed, but the
configured named of the script were kept the same, then the status
of the old script would be transferred to the new script, despite the
scripts being completely different.

This commit now checks that the script really is the same, in terms
of the path, parameters and user executing the script.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>