all: stop "unmatched quotes" warning for quoted strings
If a line with a quoted string has unbalanced quote characters
when parsed as a standard (not quoted) string, an innapropriate
warning was issued for unmatches quotes.

This commit now stops the warning.

This commit is not elegant, and it would be appreciated if a neater
solution could be found. If anyone has a better solution, please
submit a pull request or raise an issue explaining the solution.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>