all: change checking process name at reload to include not NULL checks
The code was using the reload variable as an indicator that
prev_global_data was not NULL, and this was causing some static
code analysers to to flag up NULL pointer dereferences.

The patch explicitly checks whether prev_global_data is NULL or not,
since this is synonymous with testing the reload variable.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>