vrrp: update location of iproute config files
Since iproute2 version 3.3 the location of the config files has
been configurable, with the default being /etc/iproute2. Since
version 4.4 there has been an rt_tables.d sub-directory. Version
4.10 added an rt_protos.d sub-directory, and version 6.5 added
a second directory (/usr/lib/iproute2 or /usr/lib64/iproute)
which 6.7 changed to /usr/share/iproute2 as the default.

No major distro appears to change the default locations, and the
only distro that used verion 6.5 or 6.6 was Fedora 40, but that
has now upgraded to 6.7 so we are not bothered with the /usr/lib*
options.

The two directories have configure options, and if they are not
specified, configure attempts to get the locations from the
ip-route man page or the ip executable.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>