vrrp: Don't segfault if open_sockpool_socket() fails to open sockets
If a unicast VRRP instance is configured and the unicast_src_ip does
not exist on the system, then the bind() fails and the sockets are not
opened. This commit ensures that in that case vrrp->sockets is not
dereferenced.

This is not a real fix to the problem. We need to track the addition
and removal of unicast_src_ip addresses, and enter fault state if the
address in not configured, or when it is removed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>