build: Stop _FORTITY_SOURCE redefined warnings on Ubuntu >= 24.04
Ununtu 24.04 onwards defines _FORTIFY_SOURCE=3 by default, but keepalived
was defining it to be 2, and the check for whether it was already defined
was not being done correctly.

This commit adds checking whether _FORTIFY_SOURCE is defined as a compiler
default, and now no longer explicitly sets it if it is.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>