vrrp: Restore priority 255 if duplicate address owner detected
The VRRP RFCs assume that only one device is configured as the address
owned for any VRID.

keepalived has extended functionality which detects if two (or more)
systems are configured as the address owner (this is completely
invalid configuration). To avoid multiple systems acting as address
owner, and hence all of them remaining in master mode, keepalived will
reduce an address owner's priority to 254 if the other device configured
as address owner does not go away.

This commit restores the priority of a vrrp instance to 255 if it had
reduced it to 254 to avoid multiple VRRP instances simultaneously
advertising that they are the address owner.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>