vrrp: if duplicate address owners, reduce priority if other won't
If a VRRP instance is configured as address owner and it detects another
device also advertising it is the address owner, we don't reduce our
priority if our primary IP address is higher than the other device's
primary IP address. However, if the other system, with a lower primary
IP address, won't reduce its priority (e.g. it is not a keepalived
implementation), then we will reduce our priority after a suitable time.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>