vrrp: add option for address owner to drop received VRRP packets
RFC 9568 (and RFC 5798 and RFC 3768) state that an address owner
must drop any received VRRP packets. The consequence of this is that
if there is more than one VRRP instance configured with priority 255
then they will all be in master state simultaneously. It seems more
sensible for such received packets to be processed normally, and all
but the VRRP instance with the primary IP address will revert to
backup state.

RFC 9568 appears to allow more than one instance to have priority
255, since section 8.3.2 was changed from "No more than one router
on the link is to be configured with priority 255, especially if
preemption is set" (note the contradiction here) to "only a single
VRRP Router on the link SHOULD be configured with priority 255"
and then descibes the situation if there is more than one such
router.

keepalived defaults to processing received packets when the local
priority is 255, but the option added by this patch allows working
in accordance with the RFC, i.e. to drop any received packets.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>