Fix segfault caused double erase from child_pid rb tree
In a situation when a child was timed out, but not yet processed,
the thread is THREAD_CHILD_TIMEOUT type and remains on ready queue.
If it gets terminated in this state, it needs to be removed
from rb tree child_pid and transitioned to THREAD_CHILD_TERMINATED,
but without additional moving it to ready queue as it is already there.

The erase from child_pid tree is required to clean up pid from not
terminated childs tree, but it needs to be done exactly once as rb tree
implementation is not guarded against double removal. Erasing or adding
same element multiple times, leads to malformed red-black tree and
segmentation faults.

This patch removes double erase in described scenario.

Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>