vrrp: handle a reload with no more startup_delay
During the vrrp_startup_delay time, if keepalived is reloaded with
no more startup_delay, the startup_delay is never timed out and all
received adverts is discarded.

The commit causes the startup_delay timer to be reinstated after a
reload with no more startup_delay if the timer has not yet expired.

Fixes: 58483b22c8 ("vrrp: Handle a reload before vrrp_delayed_start has expired")
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>