Skip running not idle vrrp scripts
This is a fix for a bug introduced during refactoring thread functions
in: e8e6401506c7e200cc0cb759f6d41bef89930fc7
"Thread functions return value is not used so make them void"

When a vrrp script is to be run (initially or after specified interval),
first it is checked if it's in IDLE state. If not a log message is
printed informing about skipping run due to script being either running
or timed out. However despite not being idle the code continues to run
new script process.

In heavily loaded systems this caused running multiple instances of vrrp
script at the same time.

This patch brings back missing return, which was lost during
refactoring.

Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>