vrrp: add vrrp_exit_timer_thread
Adding a vrrp_exit_timer_thread, to keep track of the fault exit timeout.
A callback is triggered which either brings up the instance if it
is triggered from the try_up_instance or just changes the state, if it
is triggered init time because the config is reloaded, while the instance
is in fault state because of a fault exit timer (Without fault init exit
delay, the instance would start in backup state).

When fault_init_exit_delay is modified, and the config is reloaded,
the new fault_init_exit_delay is applied in the following manner:
1. If the fault_init_exit_time has already passed, don't change it.
   The new fault_init_exit_delay will only be applied the next time the
   interface comes out of fault state.
2. If the new fault_init_exit_delay is greater than the old one, and
   the fault_init_exit_delay timer is presently running, increase the timer
   by the difference.
3. If the new fault_init_exit_delay is less than the old one, and the
   propagation_delay timer is in progress, reduce the timer by the
   difference between the two, provided the reduced time is greater than
   the present time.