vrrp: don't have multiple tracking objects for a VRRP instance
The code did have separate tracking objects for dynamic and non
dynamic tracking objects for a VRRP instance. It also would add
an addition dynamic tracking object every time a tracked interface
was created, causing down_instance() to be called multiple times
when an interface was deleted and previous creations of the interface.

Prior to the patch to add fault flag bits this resulted in the
vrrp instance not coming back up after the interface was recreated.

Ths issue of vrrp instances remaining in fault state after after
deletion and re-creation of interfaces is now resolved.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>